ui/pyproject.toml
[project]
name = "ui"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
